Moved duplicate code to parent class
